Update connection flow to handle connection errors and reconnect flow
